backend/requirements.txt
Django
djangorestframework
django-cors-headers
drf-spectacular
djangorestframework-simplejwt
